+++ ORBIT/manager.py
@@ -15,15 +15,18 @@ from itertools import product
 import numpy as np
 import pandas as pd
 
-from ORBIT import library
+import ORBIT
 from ORBIT.phases import DesignPhase, InstallPhase
-from ORBIT.library import initialize_library, extract_library_data
+from ORBIT.core.library import (
+    initialize_library,
+    export_library_specs,
+    extract_library_data,
+)
 from ORBIT.phases.design import (
     SparDesign,
     MonopileDesign,
     ArraySystemDesign,
     ExportSystemDesign,
-    ProjectDevelopment,
     MooringSystemDesign,
     ScourProtectionDesign,
     SemiSubmersibleDesign,
@@ -55,7 +58,6 @@ class ProjectManager:
     date_format_long = "%m/%d/%Y %H:%M"
 
     _design_phases = [
-        ProjectDevelopment,
         MonopileDesign,
         ArraySystemDesign,
         CustomArraySystemDesign,
@@ -102,17 +104,20 @@ class ProjectManager:
                 *config.get("install_phases", []),
             ],
         )
+        self._phases = {}
         self.config = self.resolve_project_capacity(config)
         self.weather = self.transform_weather_input(weather)
 
-        self.phase_starts = {}
-        self.phase_times = {}
-        self.phase_costs = {}
-        self._output_logs = []
-        self._phases = {}
-
         self.design_results = {}
         self.detailed_outputs = {}
+
+        self.system_costs = {}
+        self.installation_costs = {}
+
+        # TODO: Revise:
+        self.phase_starts = {}
+        self.phase_times = {}
+        self._output_logs = []
 
     def run_project(self, **kwargs):
         """
@@ -158,6 +163,12 @@ class ProjectManager:
 
         return self._phases
 
+    @property
+    def project_params(self):
+        """Returns defined project parameters, if found."""
+
+        return self.config.get("project_parameters", {})
+
     @classmethod
     def compile_input_dict(cls, phases):
         """
@@ -189,17 +200,27 @@ class ProjectManager:
             config = cls.merge_dicts(config, d.expected_config)
             config = cls.remove_keys(config, d.output_config)
 
-        config["commissioning"] = "float (optional, default: 0.01)"
-        config["decommissioning"] = "float (optional, default: 0.15)"
-
-        config["ncf"] = "float (optional, default: 0.4)"
-        config["offtake_price"] = "$/MWh (optional, default: 80)"
-        config["project_lifetime"] = "yrs (optional, default: 25)"
-        config["discount_rate"] = "yearly (optional, default: .025)"
-        config["opex_rate"] = "$/kW/year (optional, default: 150)"
+        config["project_parameters"] = {
+            "turbine_capex": "$/kW (optional, default: 1300)",
+            "ncf": "float (optional, default: 0.4)",
+            "offtake_price": "$/MWh (optional, default: 80)",
+            "project_lifetime": "yrs (optional, default: 25)",
+            "discount_rate": "yearly (optional, default: .025)",
+            "opex_rate": "$/kW/year (optional, default: 150)",
+            "construction_insurance": "$/kW (optional, default: 44)",
+            "construction_financing": "$/kW (optional, default: 183)",
+            "contingency": "$/kW (optional, default: 316)",
+            "commissioning": "$/kW (optional, default: 44)",
+            "decommissioning": "$/kW (optional, default: 58)",
+            "site_auction_price": "$ (optional, default: 100e6)",
+            "site_assessment_cost": "$ (optional, default: 50e6)",
+            "construction_plan_cost": "$ (optional, default: 1e6)",
+            "installation_plan_cost": "$ (optional, default: 0.25e6)",
+        }
 
         config["design_phases"] = [*design_phases.keys()]
         config["install_phases"] = [*install_phases.keys()]
+        config["orbit_version"] = str(ORBIT.__version__)
 
         return config
 
@@ -391,6 +412,19 @@ class ProjectManager:
 
         return phase_config
 
+    @property
+    def phase_ends(self):
+
+        ret = {}
+        for k, t in self.phase_times.items():
+            try:
+                ret[k] = self.phase_starts[k] + t
+
+            except KeyError:
+                pass
+
+        return ret
+
     def run_install_phase(self, name, start, **kwargs):
         """
         Compiles the phase specific configuration input dictionary for input
@@ -407,8 +441,6 @@ class ProjectManager:
         -------
         time : int | float
             Total phase time.
-        cost : int | float
-            Total phase cost.
         logs : list
             List of phase logs.
         """
@@ -422,41 +454,43 @@ class ProjectManager:
         _catch = kwargs.get("catch_exceptions", False)
         _class = self.get_phase_class(name)
         _config = self.create_config_for_phase(name)
-
-        kwargs = _config.pop("kwargs", {})
+        processes = _config.pop("processes", {})
 
         if _catch:
             try:
                 phase = _class(
-                    _config, weather=weather, phase_name=name, **kwargs
+                    _config, weather=weather, phase_name=name, **processes
                 )
                 phase.run()
 
             except Exception as e:
                 print(f"\n\t - {name}: {e}")
-                self.phase_costs[name] = e.__class__.__name__
-                self.phase_times[name] = e.__class__.__name__
-
-                return None, None, None
+                return None, None
 
         else:
-            phase = _class(_config, weather=weather, phase_name=name, **kwargs)
+            phase = _class(
+                _config, weather=weather, phase_name=name, **processes
+            )
             phase.run()
 
         self._phases[name] = phase
 
         time = phase.total_phase_time
-        cost = phase.total_phase_cost
         logs = deepcopy(phase.env.logs)
 
         self.phase_starts[name] = start
-        self.phase_costs[name] = cost
         self.phase_times[name] = time
         self.detailed_outputs = self.merge_dicts(
             self.detailed_outputs, phase.detailed_output
         )
 
-        return cost, time, logs
+        if phase.system_capex:
+            self.system_costs[name] = phase.system_capex
+
+        if phase.installation_capex:
+            self.installation_costs[name] = phase.installation_capex
+
+        return time, logs
 
     def get_phase_class(self, phase):
         """
@@ -511,8 +545,6 @@ class ProjectManager:
 
             except Exception as e:
                 print(f"\n\t - {name}: {e}")
-                self.phase_costs[name] = e.__class__.__name__
-                self.phase_times[name] = e.__class__.__name__
                 return
 
         else:
@@ -521,8 +553,6 @@ class ProjectManager:
 
         self._phases[name] = phase
 
-        self.phase_costs[name] = phase.total_phase_cost
-        self.phase_times[name] = phase.total_phase_time
         self.design_results = self.merge_dicts(
             self.design_results, phase.design_result, overwrite=False
         )
@@ -547,7 +577,7 @@ class ProjectManager:
         start = 0
 
         for name in phase_list:
-            _, time, logs = self.run_install_phase(name, start, **kwargs)
+            time, logs = self.run_install_phase(name, start, **kwargs)
 
             if logs is None:
                 continue
@@ -579,7 +609,7 @@ class ProjectManager:
         # Run defined
         for name, start in defined.items():
 
-            _, _, logs = self.run_install_phase(name, start, **kwargs)
+            _, logs = self.run_install_phase(name, start, **kwargs)
 
             if logs is None:
                 continue
@@ -625,9 +655,7 @@ class ProjectManager:
 
                 try:
                     start = self.get_dependency_start_time(target, perc)
-                    cost, time, logs = self.run_install_phase(
-                        name, start, **kwargs
-                    )
+                    _, logs = self.run_install_phase(name, start, **kwargs)
 
                     progress = True
 
@@ -805,13 +833,13 @@ class ProjectManager:
         if not self._output_logs:
             raise Exception("Project hasn't been ran yet.")
 
-        return self._output_logs
+        return sorted(self._output_logs, key=lambda l: l["time"])
 
     @property
     def project_time(self):
         """Returns total project time as the time of the last log."""
 
-        return self.project_logs[-1]["time"]
+        return self.project_actions[-1]["time"]
 
     @property
     def month_bins(self):
@@ -825,7 +853,7 @@ class ProjectManager:
         construction."""
 
         opex = self.monthly_opex
-        lifetime = self.config.get("project_lifetime", 25)
+        lifetime = self.project_params.get("project_lifetime", 25)
 
         _expense_logs = self._filter_logs(keys=["cost", "time"])
         expenses = np.array(
@@ -843,8 +871,8 @@ class ProjectManager:
     def monthly_opex(self):
         """Returns the monthly OpEx expenditures based on project size."""
 
-        rate = self.config.get("opex_rate", 150)
-        lifetime = self.config.get("project_lifetime", 25)
+        rate = self.project_params.get("opex_rate", 150)
+        lifetime = self.project_params.get("project_lifetime", 25)
 
         try:
             times, turbines = self.progress.energize_points
@@ -869,9 +897,9 @@ class ProjectManager:
         """Returns the monthly revenue based on when array system strings can
         be energized, eg. 'self.progress.energize_points'."""
 
-        ncf = self.config.get("ncf", 0.4)
-        price = self.config.get("offtake_price", 80)
-        lifetime = self.config.get("project_lifetime", 25)
+        ncf = self.project_params.get("ncf", 0.4)
+        price = self.project_params.get("offtake_price", 80)
+        lifetime = self.project_params.get("project_lifetime", 25)
 
         times, turbines = self.progress.energize_points
         dig = list(np.digitize(times, self.month_bins))
@@ -909,7 +937,7 @@ class ProjectManager:
         """Returns the net present value of the project based on
         `self.cash_flow`."""
 
-        dr = self.config.get("discount_rate", 0.025)
+        dr = self.project_params.get("discount_rate", 0.025)
         pr = (1 + dr) ** (1 / 12) - 1
 
         cash_flow = self.cash_flow
@@ -918,7 +946,7 @@ class ProjectManager:
             for i in range(1, max(cash_flow.keys()) + 1)
         ]
 
-        return self.overnight_capex - sum(_npv)
+        return (self.total_capex - self.installation_capex) - sum(_npv)
 
     @property
     def progress_logs(self):
@@ -1037,34 +1065,6 @@ class ProjectManager:
         return abs((a - b).days)
 
     @property
-    def phase_costs_per_kw(self):
-        """
-        Returns phase costs in CAPEX/kW.
-        """
-
-        _dict = {}
-        for k, capex in self.phase_costs.items():
-
-            try:
-                _dict[k] = capex / (self.capacity * 1000)
-
-            except TypeError:
-                pass
-
-        return _dict
-
-    @property
-    def overnight_capex(self):
-        """Returns the overnight capital cost of the project."""
-
-        design_phases = [p.__name__ for p in self._design_phases]
-        design_cost = sum(
-            [v for k, v in self.phase_costs.items() if k in design_phases]
-        )
-
-        return design_cost + self.turbine_capex
-
-    @property
     def overnight_capex_per_kw(self):
         """
         Returns overnight CAPEX/kW.
@@ -1079,25 +1079,32 @@ class ProjectManager:
         return capex
 
     @property
-    def installation_capex(self):
-        """
-        Returns installation related CAPEX.
-        """
+    def system_capex(self):
+        """Returns total system procurement CapEx."""
 
-        res = sum(
-            [
-                v
-                for k, v in self.phase_costs.items()
-                if k in self.config["install_phases"] and isinstance(v, Number)
-            ]
-        )
-        return res
+        return np.nansum([c for _, c in self.system_costs.items()])
+
+    @property
+    def system_capex_per_kw(self):
+        """Returns system CapEx/kW."""
+
+        try:
+            capex = self.system_capex / (self.capacity * 1000)
+
+        except TypeError:
+            capex = None
+
+        return capex
+
+    @property
+    def installation_capex(self):
+        """Returns total installation related CapEx."""
+
+        return np.nansum([c for _, c in self.installation_costs.items()])
 
     @property
     def installation_capex_per_kw(self):
-        """
-        Returns installation related CAPEX/kW.
-        """
+        """Returns installation CapEx/kW."""
 
         try:
             capex = self.installation_capex / (self.capacity * 1000)
@@ -1109,83 +1116,16 @@ class ProjectManager:
 
     @property
     def bos_capex(self):
-        """
-        Returns BOS CAPEX not including commissioning and decommissioning.
-        """
+        """Returns total balance of system CapEx."""
 
-        return sum([v for _, v in self.phase_costs.items()])
+        return self.system_capex + self.installation_capex
 
     @property
     def bos_capex_per_kw(self):
-        """
-        Returns BOS CAPEX/kW not including commissioning and decommissioning.
-        """
+        """Returns balance of system CapEx/kW."""
 
         try:
             capex = self.bos_capex / (self.capacity * 1000)
-
-        except TypeError:
-            capex = None
-
-        return capex
-
-    @property
-    def commissioning(self):
-        """
-        Returns the cost of commissioning based on the configured phases.
-        Defaults to 1% of total BOS CAPEX.
-        """
-
-        _comm = self.config.get("commissioning", 0.0)
-        if (_comm < 0.0) or (_comm > 1.0):
-            raise ValueError("'commissioning' must be between 0 and 1")
-
-        total = self.bos_capex + self.turbine_capex
-
-        comm = total * _comm
-        return comm
-
-    @property
-    def commissioning_per_kw(self):
-        """
-        Returns the cost of commissioning per kW.
-        """
-
-        try:
-            capex = self.commissioning / (self.capacity * 1000)
-
-        except TypeError:
-            capex = None
-
-        return capex
-
-    @property
-    def decommissioning(self):
-        """
-        Returns the cost of decommissioning based on the configured
-        installation phases. Defaults to 15% of installation CAPEX.
-        """
-
-        _decomm = self.config.get("decommissioning", 0.0)
-        if (_decomm < 0.0) or (_decomm > 1.0):
-            raise ValueError("'decommissioning' must be between 0 and 1")
-
-        try:
-            decomm = self.installation_capex * _decomm
-
-        except KeyError:
-            return 0.0
-
-        return decomm
-
-    @property
-    def decommissioning_per_kw(self):
-        """
-        Returns the cost of decommissioning per kW.
-        """
-
-        try:
-            capex = self.decommissioning / (self.capacity * 1000)
 
         except TypeError:
             capex = None
@@ -1198,49 +1138,109 @@ class ProjectManager:
         Returns the total turbine CAPEX.
         """
 
-        _capex = self.config.get("turbine_capex", 0.0)
+        _capex = self.project_params.get("turbine_capex", 1300)
         try:
             num_turbines = self.config["plant"]["num_turbines"]
             rating = self.config["turbine"]["turbine_rating"]
 
         except KeyError:
-            print(
-                f"Turbine CAPEX not included in commissioning. Required "
+            raise KeyError(
+                f"Total turbine CAPEX can't be calculated. Required "
                 f"parameters 'plant.num_turbines' or 'turbine.turbine_rating' "
                 f"not found."
             )
-            return 0.0
 
         capex = _capex * num_turbines * rating * 1000
         return capex
 
     @property
     def turbine_capex_per_kw(self):
-        """
-        Returns the turbine CAPEX/kW.
-        """
+        """Returns the turbine CapEx/kW."""
 
-        _capex = self.config.get("turbine_capex", None)
+        _capex = self.project_params.get("turbine_capex", None)
         return _capex
 
     @property
-    def total_capex(self):
+    def overnight_capex(self):
+        """Returns the overnight capital cost of the project."""
+
+        return self.system_capex + self.turbine_capex
+
+    @property
+    def soft_capex(self):
+        """Returns total project cost costs."""
+
+        try:
+            capex = self.soft_capex_per_kw * self.capacity * 1000
+
+        except TypeError:
+            capex = None
+
+        return capex
+
+    @property
+    def soft_capex_per_kw(self):
         """
-        Returns total project CAPEX including commissioning and decommissioning.
+        Returns project soft costs per kW. Default numbers are based on the
+        Cost of Energy Review (Stehly and Beiter 2018).
         """
 
-        return (
-            self.bos_capex
-            + self.turbine_capex
-            + self.commissioning
-            + self.decommissioning
+        insurance = self.project_params.get("construction_insurance", 44)
+        financing = self.project_params.get("construction_financing", 183)
+        contingency = self.project_params.get("contingency", 316)
+        commissioning = self.project_params.get("commissioning", 44)
+        decommissioning = self.project_params.get("decommissioning", 58)
+
+        return sum(
+            [insurance, financing, contingency, commissioning, decommissioning]
         )
 
     @property
+    def project_capex(self):
+        """
+        Returns project related CapEx line items. To override the defaults,
+        the keys below should be passed to the 'project_parameters' subdict.
+        """
+
+        site_auction = self.project_params.get("site_auction_price", 100e6)
+        site_assessment = self.project_params.get("site_assessment_cost", 50e6)
+        construction_plan = self.project_params.get(
+            "construction_plan_cost", 1e6
+        )
+        installation_plan = self.project_params.get(
+            "installation_plan_cost", 0.25e6
+        )
+
+        return sum(
+            [
+                site_auction,
+                site_assessment,
+                construction_plan,
+                installation_plan,
+            ]
+        )
+
+    @property
+    def project_capex_per_kw(self):
+        """Returns project related CapEx per kW."""
+
+        try:
+            capex = self.project_capex / (self.capacity * 1000)
+
+        except TypeError:
+            capex = None
+
+        return capex
+
+    @property
+    def total_capex(self):
+        """Returns total project CapEx including soft costs."""
+
+        return self.bos_capex + self.turbine_capex + self.soft_capex
+
+    @property
     def total_capex_per_kw(self):
-        """
-        Returns total BOS CAPEX/kW including commissioning and decommissioning.
-        """
+        """Returns total CapEx/kW."""
 
         try:
             capex = self.total_capex / (self.capacity * 1000)
@@ -1261,7 +1261,7 @@ class ProjectManager:
             Name to use for the file.
         """
 
-        library.export_library_specs("config", file_name, self.config)
+        export_library_specs("config", file_name, self.config)
 
 
 class ProjectProgress: