+++ ORBIT/phases/install/oss_install/standard.py
@@ -33,7 +33,7 @@ class OffshoreSubstationInstallation(InstallPhase):
     expected_config = {
         "num_substations": "int",
         "oss_install_vessel": "dict | str",
-        "num_feeders": "int",
+        "num_feeders": "int (optional, default: 1)",
         "feeder": "dict | str",
         "site": {"distance": "km", "depth": "m"},
         "port": {
@@ -41,12 +41,17 @@ class OffshoreSubstationInstallation(InstallPhase):
             "monthly_rate": "USD/mo (optional)",
             "name": "str (optional)",
         },
-        "offshore_substation_topside": {"deck_space": "m2", "mass": "t"},
+        "offshore_substation_topside": {
+            "deck_space": "m2",
+            "mass": "t",
+            "unit_cost": "USD",
+        },
         "offshore_substation_substructure": {
             "type": "Monopile",
             "deck_space": "m2",
             "mass": "t",
             "length": "m",
+            "unit_cost": "USD",
         },
     }
 
@@ -66,10 +71,18 @@ class OffshoreSubstationInstallation(InstallPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
 
         self.initialize_port()
         self.setup_simulation(**kwargs)
+
+    @property
+    def system_capex(self):
+        """Returns procurement CapEx of the offshore substations."""
+
+        return self.config["num_substations"] * (
+            self.config["offshore_substation_topside"]["unit_cost"]
+            + self.config["offshore_substation_substructure"]["unit_cost"]
+        )
 
     def setup_simulation(self, **kwargs):
         """
@@ -143,7 +156,7 @@ class OffshoreSubstationInstallation(InstallPhase):
         Initializes feeder barge objects.
         """
 
-        number = self.config.get("num_feeders", None)
+        number = self.config.get("num_feeders", 1)
         feeder_specs = self.config.get("feeder", None)
 
         self.feeders = []