+++ ORBIT/phases/install/quayside_assembly_tow/moored.py
@@ -35,6 +35,7 @@ class MooredSubInstallation(InstallPhase):
         "substructure": {
             "takt_time": "int | float (optional, default: 0)",
             "towing_speed": "int | float (optional, default: 6 km/h)",
+            "unit_cost": "USD",
         },
         "site": {"depth": "m", "distance": "km"},
         "plant": {"num_turbines": "int"},
@@ -65,7 +66,6 @@ class MooredSubInstallation(InstallPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
 
         self.setup_simulation(**kwargs)
 
@@ -85,6 +85,12 @@ class MooredSubInstallation(InstallPhase):
         self.initialize_queue()
         self.initialize_towing_groups()
         self.initialize_support_vessel()
+
+    @property
+    def system_capex(self):
+        """Returns total procurement cost of the substructures."""
+
+        return self.num_turbines * self.config["substructure"]["unit_cost"]
 
     def initialize_substructure_production(self):
         """