+++ ORBIT/phases/install/scour_protection_install/standard.py
@@ -12,7 +12,7 @@ import simpy
 from marmot import process
 
 from ORBIT.core import Vessel
-from ORBIT.core._defaults import process_times as pt
+from ORBIT.core.defaults import process_times as pt
 from ORBIT.phases.install import InstallPhase
 from ORBIT.core.exceptions import CargoMassExceeded, InsufficientAmount
 
@@ -34,7 +34,10 @@ class ScourProtectionInstallation(InstallPhase):
             "monthly_rate": "USD/mo (optional)",
             "name": "str (optional)",
         },
-        "scour_protection": {"tons_per_substructure": "float"},
+        "scour_protection": {
+            "tonnes_per_substructure": "t",
+            "cost_per_tonne": "USD/t",
+        },
     }
 
     phase = "Scour Protection Installation"
@@ -55,7 +58,6 @@ class ScourProtectionInstallation(InstallPhase):
 
         config = self.initialize_library(config, **kwargs)
         self.config = self.validate_config(config)
-        self.extract_defaults()
 
         self.setup_simulation(**kwargs)
 
@@ -82,9 +84,11 @@ class ScourProtectionInstallation(InstallPhase):
                 / 1000.0
             )
 
-        self.tons_per_substructure = ceil(
-            self.config["scour_protection"]["tons_per_substructure"]
+        self.tonnes_per_substructure = ceil(
+            self.config["scour_protection"]["tonnes_per_substructure"]
         )
+
+        self.cost_per_tonne = self.config["scour_protection"]["cost_per_tonne"]
 
         install_scour_protection(
             self.spi_vessel,
@@ -92,8 +96,18 @@ class ScourProtectionInstallation(InstallPhase):
             site_distance=site_distance,
             turbines=self.num_turbines,
             turbine_distance=turbine_distance,
-            tons_per_substructure=self.tons_per_substructure,
+            tonnes_per_substructure=self.tonnes_per_substructure,
             **kwargs,
+        )
+
+    @property
+    def system_capex(self):
+        """Returns total procurement cost of scour protection material."""
+
+        return (
+            self.num_turbines
+            * self.tonnes_per_substructure
+            * self.cost_per_tonne
         )
 
     def initialize_port(self):
@@ -136,7 +150,7 @@ def install_scour_protection(
     site_distance,
     turbines,
     turbine_distance,
-    tons_per_substructure,
+    tonnes_per_substructure,
     **kwargs,
 ):
     """
@@ -155,8 +169,8 @@ def install_scour_protection(
         For now this assumes it traverses an edge and not a diagonal.
     turbines_to_install : int
         Number of turbines where scouring protection must be installed.
-    tons_per_substructure : int
-        Number of tons required to be installed at each substation
+    tonnes_per_substructure : int
+        Number of tonnes required to be installed at each substation
     """
 
     while turbines > 0:
@@ -172,14 +186,14 @@ def install_scour_protection(
             vessel.at_site = True
 
         elif vessel.at_site:
-            if vessel.rock_storage.level >= tons_per_substructure:
+            if vessel.rock_storage.level >= tonnes_per_substructure:
                 # Drop scour protection material
-                yield drop_material(vessel, tons_per_substructure, **kwargs)
+                yield drop_material(vessel, tonnes_per_substructure, **kwargs)
                 turbines -= 1
 
                 # Transit to another turbine
                 if (
-                    vessel.rock_storage.level >= tons_per_substructure
+                    vessel.rock_storage.level >= tonnes_per_substructure
                     and turbines > 0
                 ):
                     yield vessel.transit(turbine_distance)