+++ tests/phases/design/test_spar_design.py
@@ -51,7 +51,7 @@ def test_design_kwargs():
 
     s = SparDesign(base)
     s.run()
-    base_cost = s.total_phase_cost
+    base_cost = s.total_cost
 
     for k, v in test_kwargs.items():
 
@@ -61,6 +61,6 @@ def test_design_kwargs():
 
         s = SparDesign(config)
         s.run()
-        cost = s.total_phase_cost
+        cost = s.total_cost
 
         assert cost != base_cost